+++ crates/core/src/cache/module_cache.rs
@@ -35,6 +35,7 @@ pub struct CachedModule {
   pub dependencies: Vec<CachedModuleDependency>,
   pub package_name: String,
   pub package_version: String,
+  pub entry_name: Option<String>,
 }
 
 impl CachedModule {