+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @farmfe/core
 
+## 0.14.2
+
+### Patch Changes
+
+- cf0dc914: Fix cjs export not found
+- c1a4fcc8: fix #747
+- Updated dependencies [c1a4fcc8]
+  - @farmfe/runtime@0.8.2
+
 ## 0.14.1
 
 ### Patch Changes